split config and apply a review
